AJUSTES PARA ACEITAR QUANTIDADE NULA VINDA DO INICIAL

+++ Producao/Views/kit/manutencao/ViewDetalhesKitManutencao.xaml.cs
@@ -471,7 +471,7 @@ namespace Producao.Views.kit.solucao
                     obs = dado?.obs,
                     orient_montagem = dado?.orient_montagem,
                     orient_desmont = dado?.orient_desmont,
-                    qtd = dado.qtd,
+                    qtd = (double)dado.qtd,
                     alterado_por = Environment.UserName,
                     alterado_em = DateTime.Now
                 //ordem = dado?.id
@@ -512,7 +512,7 @@ namespace Producao.Views.kit.solucao
                     incluidopordesc = vm?.CheckListGeral?.incluidopordesc,
                     kp = vm?.CheckListGeral?.kp,
                     orient_desmont = vm?.CheckListGeral?.orient_desmont,
-                    qtd = vm.CheckListGeral.qtd,
+                    qtd = (double)vm.CheckListGeral.qtd,
                     coduniadicional = vm?.CheckListGeral?.coduniadicional,
                     codcompl = vm?.CheckListGeral?.codcompl,
                     nivel = vm?.CheckListGeral?.nivel,
